refactor(gsd-db): peel manifest restore + legacy import into db/writers/import-restore.ts
Phase A, commit 4. Extracts restoreManifest (~228 lines) and
bulkInsertLegacyHierarchy + Legacy* payload interfaces into
db/writers/import-restore.ts. Both are self-contained bulk operations that own
their own transaction().

Also consolidates the Single Writer Layer re-exports (memory, reconcile,
import-restore) into one stable block right after the engine re-export, instead
of appended at the file tail where a later tail extraction could truncate them.
Drops the now-unused StateManifest type import from gsd-db.ts.

gsd-db.ts: 2,158 → 1,853 lines.

Verified: typecheck:extensions clean; single-writer-invariant (incl. the
expected-wrappers export check) + compaction-snapshot (restoreManifest e2e) +
collect-from-manifest + manifest-status green (21 tests).
